Create Phase II Python Code Draft-Hannah
Hey Guys,
I am working on the phase II portion of the project and I am getting stuck on the inertia.  I don't remember learning about this in class and I am not sure how to calculate it.  Above is the code that I have developed so far.  I will continue to research inertias and how to calculate them.